+++ src/components/molecules/generated-image/GeneratedImage.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import CloseButton from '@/components/atoms/close-button/CloseButton';
+import CloseButton from '@/components/atoms/button/CloseButton';
 
 interface Props {
   key: number;
@@ -12,7 +12,7 @@ export default function GeneratedImage({ key, url, onClose }: Props) {
       <div className="relative">
         <CloseButton id={key} onClose={() => onClose(key)} />
         {url && (
-          <div className="relative overflow-hidden rounded-[24px] max-xl:size-[206px] max-sm:size-[167px] xl:size-[180px]">
+          <div className="max-xl:size-[206px] max-sm:size-[167px] relative overflow-hidden rounded-[24px] xl:size-[180px]">
             <Image fill src={url} alt="이미지 생성" className="object-cover" />
           </div>
         )}
